test: bonjour e2e
